Moved some helper components for visits to visits/helpers
