+++ src/visits/VisitsStats.js
@@ -9,8 +9,8 @@ import DateRangeRow from '../utils/DateRangeRow';
 import Message from '../utils/Message';
 import { formatDate } from '../utils/helpers/date';
 import { useToggle } from '../utils/helpers/hooks';
-import SortableBarGraph from './SortableBarGraph';
-import GraphCard from './GraphCard';
+import SortableBarGraph from './helpers/SortableBarGraph';
+import GraphCard from './helpers/GraphCard';
 import LineChartCard from './helpers/LineChartCard';
 import VisitsTable from './VisitsTable';
 import { VisitsInfoType } from './types';
