+++ src/visits/helpers/GraphCard.scss
